requirements.txt
Django~=3.2.6
djangorestframework~=3.12.4
knox~=0.1.14
django-filter~=2.4.0